Change basemap URL to Mapbox style
I think this is the only piece that needs to change in this file, with the rest (including Mapbox access token) needing changes in ResourceMap.vue - but I'm not familiar with this code structure! There are other references to Carto in here that I'm not sure if it's going to cause problems or if that's just about the database.